on submit changed to on click for form submisison
using 'submit' listener on contact form to initiate ajax call
was causing an error if the form had to be submitted more than once (for
instance because of form validation or recaptcha error). Changing to
using a 'click' listener on the form submit button seems to have
resolved this.

Also added some console.log statements to help debug the problems we are
having on the deployed instances. These will be removed once the issus
is resolved.